requirements.txt
Django==4.2.14
django-comment-migrate==0.1.7
django-cors-headers==4.4.0
django-filter==24.2
django-ranged-response==0.2.0
djangorestframework==3.15.2
django-restql==0.15.4
django-simple-captcha==0.6.0
django-timezone-field==7.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.7
PyMySQL>=1.0.0
pypinyin==0.51.0
ua-parser==0.18.0
pyparsing==3.1.2
openpyxl==3.1.5
requests==2.32.4
typing-extensions==4.12.2
tzlocal==5.2
channels==4.1.0
channels-redis==4.2.0
websockets>=12.0
wsproto>=1.2.0
user-agents==2.2.0
six==1.16.0
whitenoise==6.7.0
psycopg2-binary
uvicorn==0.30.3
gunicorn==23.0.0
gevent>=24.2.1
Pillow==10.4.0
pyinstaller>=6.10.0
dvadmin3-celery==3.1.6
oss2==2.19.1
cos-python-sdk-v5==1.9.37
python-docx==1.1.2
PyMuPDF==1.24.7
beautifulsoup4==4.12.3
mammoth==1.6.0
bleach==6.1.0
lxml==4.9.3
langgraph-sdk>=0.1.0
PyJWT==2.8.0
# RSA解密库
pycryptodome>=3.19.0
# 腾讯云SDK
tencentcloud-sdk-python>=3.0.1000
